Parse a map from the ORB_SLAM2
